requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
pytz==2023.3
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
